+++ payrollManagementSystem/src/main/java/com/payrollManagementSystem/service/OtpService.java
@@ -5,7 +5,7 @@ import com.payrollManagementSystem.entity.Otp;
 public interface OtpService {
 	void addOtpEntry(Otp otpObject);
 
-	Otp getMappedOtp(int employeeId);
+	Otp getMappedOtp(long employeeId);
 
 	boolean validateOtp(Otp otp);
